+++ trunk/conn.h
@@ -42,9 +42,9 @@ typedef struct
 	
 	ftp_t ftp[1];
 	http_t http[1];
-	int size;		/* File size, not 'connection size'..	*/
-	int currentbyte;
-	int lastbyte;
+	long long size;		/* File size, not 'connection size'..	*/
+	long long currentbyte;
+	long long lastbyte;
 	int fd;
 	int enabled;
 	int supported;
