+++ trunk/http.h
@@ -33,8 +33,8 @@ typedef struct
 	char headers[MAX_QUERY];
 	int proto;			/* FTP through HTTP proxies	*/
 	int proxy;
-	int firstbyte;
-	int lastbyte;
+	long long int firstbyte;
+	long long int lastbyte;
 	int status;
 	int fd;
 	char *local_if;
@@ -46,6 +46,6 @@ void http_get( http_t *conn, char *lurl );
 void http_addheader( http_t *conn, char *format, ... );
 int http_exec( http_t *conn );
 char *http_header( http_t *conn, char *header );
-int http_size( http_t *conn );
+long long int http_size( http_t *conn );
 void http_encode( char *s );
 void http_decode( char *s );
